Trait fj_wasm_shim::host::Host

pub trait Host: Sized {
    fn print(&mut self, message: &str);
    fn abort(&mut self, message: &str);
    fn time(&mut self) -> Timespec;
}

Required Methods

fn print(&mut self, message: &str)

Write something to the console without appending a newline.

fn abort(&mut self, message: &str)

Immediately halt execution with a message.

fn time(&mut self) -> Timespec

The amount of time that has passed since the UNIX epoch (UTC).